Added build directory to seperate auto updating source from running code
- Added pipx to provide global hagrid editable usage
- Small fixes to HAGrid
- Logging cron output

+++ packages/grid/scripts/cron.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
+
+# cronjob logs: $ tail -f /var/log/syslog | grep -i cron
+
 # $1 is the PySyft dir
 # $2 is the git repo like: https://github.com/OpenMined/PySyft
 # $3 is the branch like: dev
 # $4 is the permission user like: om
 # $5 is the permission group like: om
 # $6 is the node type like: domain
-# #7 is the node name like: node
+# $7 is the node name like: node
+# $8 is the build directory where we copy the source so we dont trigger hot reloading
 
 # these commands cant be used because they trigger hot reloading
 # however without them accidental changes to the working tree might cause issues
@@ -18,15 +22,32 @@
 
 cd $1
 START_HASH=$(git rev-parse HEAD)
-git remote rm origin || true
-git remote add origin https://github.com/$2
-git fetch origin
-git checkout $3
-git pull origin $3 --rebase
-chown -R $4:$5 .
+CURRENT_REMOTE=$(git remote -v | head -n 1 | cut -d ' ' -f 1 | awk '{print $2}')
+CURRENT_BRANCH=$(git branch --show-current)
+echo "Running autoupdate CRON"
+if [ "$CURRENT_REMOTE" != "$2" ]; then
+    echo "Switching remotes to: ${2}"
+    git remote rm origin || true
+    git remote add origin https://github.com/$2
+    git fetch origin
+    echo "Checking out branch: ${3}"
+    git checkout $3 --force
+    git pull origin $3 --rebase
+    chown -R $4:$5 .
+    elif [ "$CURRENT_BRANCH" != "$3" ]; then
+    echo "Checking out branch: ${3}"
+    git checkout $3 --force
+    git pull origin $3 --rebase
+    chown -R $4:$5 .
+fi
+
 END_HASH=$(git rev-parse HEAD)
 
 if [ "$START_HASH" != "$END_HASH" ]; then
-    echo "Code has changed to running hagrid"
-    ANSIBLE_CONFIG=/home/om/PySyft/packages/grid/ansible.cfg ansible-playbook --connection=local -i 127.0.0.1, /home/om/PySyft/packages/grid/ansible/site.yml -e "node_type=$6 node_name=$7 github_repo=$2 repo_branch=$3 deploy_only=true"
+    echo "Code has changed so redeploying with HAGrid"
+    rm -rf ${8}
+    cp -r ${1} ${8}
+    chown -R $4:$5 ${8}
+    runuser -l ${4} -c 'hagrid launch ${7} ${6} to localhost --repo=${2} --branch=${3}'
 fi
+echo "Finished autoupdate CRON"
